+++ packages/api-v4/src/account/account.ts
@@ -1,5 +1,4 @@
 import {
-  SendCodeToPhoneNumberSchema,
   updateAccountSchema,
   UpdateAccountSettingsSchema,
 } from '@linode/validation/lib/account.schema';
@@ -12,8 +11,6 @@ import {
   CancelAccountPayload,
   NetworkUtilization,
   Agreements,
-  SendPhoneVerificationCodePayload,
-  VerifyVerificationCodePayload,
 } from './types';
 
 /**
@@ -110,34 +107,6 @@ export const getAccountAgreements = () =>
 export const signAgreement = (data: Partial<Agreements>) => {
   return Request<{}>(
     setURL(`${BETA_API_ROOT}/account/agreements`),
-    setMethod('POST'),
-    setData(data)
-  );
-};
-
-/**
- * sendCodeToPhoneNumber
- *
- * Sends a one-time password via SMS to be used to verify a phone number.
- */
-export const sendCodeToPhoneNumber = (
-  data: SendPhoneVerificationCodePayload
-) => {
-  return Request<{}>(
-    setURL(`${API_ROOT}/account/phone-number`),
-    setMethod('POST'),
-    setData(data, SendCodeToPhoneNumberSchema)
-  );
-};
-
-/**
- * verifyPhoneNumberCode
- *
- * Verifies a one-time password sent using `sendCodeToPhoneNumber`.
- */
-export const verifyPhoneNumberCode = (data: VerifyVerificationCodePayload) => {
-  return Request<{}>(
-    setURL(`${API_ROOT}/account/phone-number/verify`),
     setMethod('POST'),
     setData(data)
   );
